separate type checking from build script
