+++ js/node/node-backend.js
@@ -1,6 +1,9 @@
 import { isMainThread, parentPort, workerData, Worker } from 'worker_threads';
 import os from 'os';
-import wasm_ from '../../../_node_bindings/plonk_wasm.cjs';
+// TODO: the modified directory _node_bindings (with underscore) is a hack to
+// prevent typescript from processing this folder.
+// any cleaner solution is appreciated!
+import wasm_ from '../../compiled/_node_bindings/plonk_wasm.cjs';
 import { fileURLToPath } from 'url';
 let url = import.meta.url;
 let filename = url !== undefined ? fileURLToPath(url) : __filename;
