+++ js/wrapper.js
@@ -1,5 +1,8 @@
 import { wasm, withThreadPool } from './node/node-backend.js';
-import { default as snarky } from '../../_node_bindings/snarky_js_node.bc.cjs';
+// TODO: the modified directory _node_bindings (with underscore) is a hack to
+// prevent typescript from processing this folder.
+// any cleaner solution is appreciated!
+import { default as snarky } from '../compiled/_node_bindings/snarky_js_node.bc.cjs';
 
 export { getSnarky, getWasm, withThreadPool };
